#22 Background, Cleanup and bounds

+++ core/src/test/java/com/rocketracer/game/ECS/Components/FuelComponentTest.java
@@ -21,7 +21,7 @@ public class FuelComponentTest {
     }
     @Test
     public void testFuelLevel() {
-        float fuelLevel = 100;
+        int fuelLevel = 100;
         assertEquals(fuelLevel, fuelComponent.fuelLevel);
     }
 }
